+++ aws/terraform/provider.go
@@ -41,6 +41,24 @@ func (p *Provider) ResourceComponents(tfRes terraform.Resource) []query.Componen
 			return nil
 		}
 		return p.newVolume(vals).Components()
+	case "aws_elasticache_cluster":
+		vals, err := decodeElastiCacheValues(tfRes.Values)
+		if err != nil {
+			return nil
+		}
+		return p.newElastiCache(vals).Components()
+	case "aws_elasticache_replication_group":
+		vals, err := decodeElastiCacheReplicationValues(tfRes.Values)
+		if err != nil {
+			return nil
+		}
+		return p.newElastiCacheReplication(vals).Components()
+	case "aws_eip":
+		vals, err := decodeElasticIPValues(tfRes.Values)
+		if err != nil {
+			return nil
+		}
+		return p.newElasticIP(vals).Components()
 	case "aws_db_instance":
 		vals, err := decodeDBInstanceValues(tfRes.Values)
 		if err != nil {